feat(driver-mongodb): split tests
