+++ test/datacap-test-driver/src/test/java/io/edurt/datacap/test/mongo/MongoJdbcDriverAggregationTest.java
@@ -1,0 +1,44 @@
+package io.edurt.datacap.test.mongo;
+
+import lombok.extern.slf4j.Slf4j;
+import org.junit.Test;
+
+import java.sql.ResultSet;
+import java.sql.SQLException;
+
+import static org.junit.Assert.assertTrue;
+
+@Slf4j
+public class MongoJdbcDriverAggregationTest
+        extends MongoJdbcBaseTest
+{
+    @Test
+    public void testOrderBy()
+            throws SQLException
+    {
+        log.info("Test order by");
+        try (ResultSet rs = statement.executeQuery("SELECT * FROM test.sample ORDER BY name DESC, value DESC")) {
+            assertTrue(rs.next());
+        }
+    }
+
+    @Test
+    public void testLimit()
+            throws SQLException
+    {
+        log.info("Test limit");
+        try (ResultSet rs = statement.executeQuery("SELECT * FROM test.sample LIMIT 1")) {
+            assertTrue(rs.next());
+        }
+    }
+
+    @Test
+    public void testGroupBy()
+            throws SQLException
+    {
+        log.info("Test group by");
+        try (ResultSet rs = statement.executeQuery("SELECT name FROM test.sample GROUP BY name")) {
+            assertTrue(rs.next());
+        }
+    }
+}
